feat: add navigation links and components to NavLinks
Navigation links and components were created for each of the NavBar's
display configurations (Mobile, Desktop and the Offcanvas element).
Components were used where current user information was required (e.g
links to the profile).

Conditional logic was altered in the NavBar to render variables and/or
components based on the current users authentication state.