fix: TypeError when handleLogout called
Calling handleLogout caused a TypeError to be thrown,  which was traced
to a mistake in the way the UserContext was being set.

To resolve this the I needed to use the SetCurrentUserContext which
involved refactoring the handleLogout function into a custom hook that
returned a callback function as not to violate the Rules of Hooks. This
function can then be invoked when I needed to log a user out.