fix: resolve update on unmounted component error
Problem: Calling the handleSubmit function when submitting the ProfileEditForm,
updates the API with a PUT request, updates the CurrentUserContext
(setCurrentUser), and redirects the user back to their Profile.

currentUser (as consumed from the CurrentUserContext) had been set in
the useEffect hook dependency array, which caused the useEffect hook to
execute again and try to perform a state update on the unmounted form
component.

Solution: Resolved by removing 'currentUser' from the useEffect hook
dependency array.