fix: add error handling for 400 errors when fetching property
When a property page component is mounted it tries to request the
property and the authenticated users notes for the property, using the
Promise.all method.

The 404 error returned when fetching a property that doesn't exist had
been handled. As the promises can be returned in any order, the error
checking has been updated to account for the failure of the notes
request (which returns a 400 Bad Request error with the message "Select
a valid choice. That choice is not one of the available choices.").